linux: Move `tls.h` API to `src/new`
As part of this, `TLS_INFO_TX_MAX_PAYLOAD_LEN` and `TLS_INFO_MAX` are
added.

Link: https://github.com/torvalds/linux/blob/ab9de95c9cf952332ab79453b4b5d1bfca8e514f/include/uapi/linux/tls.h
