+++ libc-test/semver/linux.txt
@@ -3665,9 +3665,11 @@ TLS_CONF_HW_RECORD
 TLS_CONF_SW
 TLS_GET_RECORD_TYPE
 TLS_INFO_CIPHER
+TLS_INFO_MAX
 TLS_INFO_RXCONF
 TLS_INFO_RX_NO_PAD
 TLS_INFO_TXCONF
+TLS_INFO_TX_MAX_PAYLOAD_LEN
 TLS_INFO_UNSPEC
 TLS_INFO_VERSION
 TLS_INFO_ZC_RO_TX
