+++ src/new/linux_uapi/linux/mod.rs
@@ -11,3 +11,4 @@ pub(crate) mod membarrier;
 pub(crate) mod mount;
 pub(crate) mod netlink;
 pub(crate) mod pidfd;
+pub(crate) mod tls;
